Add TODO for editor content error checks
Either possible by reimplementing content errors in the editor itself, or by sending off the new doc to an API service and reusing the existing content error code. Would be nice to have so content editors get instant feedback of whether the content is valid or not, and don't have to wait around for ETL to load the new commit.